Created Ceph backend
Added the Ceph backend handler. This handler allows you to pass
a raw image filename, metadata, and a Ceph pool name. It will take
this data and load the raw data into the Ceph cluster. However, it will
create the RBD with the same name as the raw image file, not a UUID.